+++ BenchmarkTool/Generators/IRecord.cs
@@ -5,12 +5,12 @@ namespace BenchmarkTool.Generators
     public interface IRecord
     {
         int SensorID { get; set; }
-        // float Value { get; set; }
-        float[] ValuesArray { get; set; }  
+        // double Value { get; set; }
+        double[] ValuesArray { get; set; }  
         DateTime Time { get; set; }
         // bool polyDim { get;  }
 
-        float getFirstValue(){
+        double getFirstValue(){
             return ValuesArray[0];
         }
