+++ BenchmarkTool/Generators/RecordClickhouse.cs
@@ -6,29 +6,29 @@ namespace BenchmarkTool.Generators
     public class RecordClickhouse : IRecord, IEnumerable
     {
         public int SensorID { get; set; }
-        // public float Value { get; set; }
-        public float[] ValuesArray { get; set; }
+        // public double Value { get; set; }
+        public double[] ValuesArray { get; set; }
         public DateTime Time { get; set; }
 
-        public RecordClickhouse(int sensorId, DateTime timestamp, float value)
+        public RecordClickhouse(int sensorId, DateTime timestamp, double value)
         {
             SensorID = sensorId;
             Time = timestamp;
             // Value = value;
-                        ValuesArray = new float[1];
+                        ValuesArray = new double[1];
 
             ValuesArray[0] = value;
 
         }
         public bool polyDim { get; }
 
-        public RecordClickhouse(int sensorId, DateTime timestamp, float[] values)
+        public RecordClickhouse(int sensorId, DateTime timestamp, double[] values)
         {
             SensorID = sensorId;
             Time = timestamp;
             ValuesArray = values;
         }
-        float getFirstValue()
+        double getFirstValue()
         {
             return ValuesArray[0];
         }
