+++ BenchmarkTool/Generators/RecordMySQLDB.cs
@@ -6,29 +6,29 @@ namespace BenchmarkTool.Generators
     public class RecordMySQLDB : IRecord
     {
         public int SensorID { get; set; }
-        // public float Value { get; set; } TODO clean
-        public float[] ValuesArray { get; set; }
+        // public double Value { get; set; } TODO clean
+        public double[] ValuesArray { get; set; }
         // public bool polyDim { get;  }
         public DateTime Time { get; set; }
 
-        public RecordMySQLDB(int sensorId, DateTime timestamp, float value)
+        public RecordMySQLDB(int sensorId, DateTime timestamp, double value)
         {
             SensorID = sensorId;
             Time = timestamp;
             // Value = value;
-                        ValuesArray = new float[1];
+                        ValuesArray = new double[1];
 
             ValuesArray[0] = value;
 
         }
-        public RecordMySQLDB(int sensorId, DateTime timestamp, float[] values)
+        public RecordMySQLDB(int sensorId, DateTime timestamp, double[] values)
         {
             SensorID = sensorId;
             Time = timestamp;
             ValuesArray = values;
         }
 
-        float getFirstValue(){
+        double getFirstValue(){
             return ValuesArray[0];
         }
         public IEnumerator GetEnumerator()
